+++ _13_30day/src/Person.java
@@ -1,0 +1,13 @@
+package src;
+
+public class Person {
+
+    protected String firstName;
+    protected String lastName;
+    int idNumber;
+
+    public void printPerson(){
+        System.out.println("Name: " + lastName + ", " + firstName + "\nID: " + idNumber);
+    }
+
+}
